+++ shop/src/main/java/com/shop/GH/AdminMember.java
@@ -9,6 +9,7 @@ import javax.inject.Inject;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -19,20 +20,24 @@ import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.servlet.ModelAndView;
 
 import com.shop.GH.service.MemberService;
+import com.shop.GH.service.admin.ShopService;
 import com.shop.GH.vo.MemberVO;
-import com.shop.GH.vo.ShopAddVO;
+import com.shop.GH.vo.admin.ShopVO;
 
 
 /**
  * @author 김건수
- * @관리자 컨트롤러
+ * @관리자 멤버 컨트롤러
  *
  */
 @Controller
-public class AdminController {
+public class AdminMember {
 	
 	@Inject
 	private MemberService m_service;
+	
+	@Autowired
+	private ShopService shop;
 	
 //	@RequestMapping(value = "admin/home", method = RequestMethod.GET)
 //	public ModelAndView homess() {							
@@ -54,63 +59,6 @@ public class AdminController {
 		}
 
 	
-	
-	
-	
-	
-	
-	/**
-	 * @author 김건수
-	 * @여기부터 상품쪽
-	 */
-	
-	//상품등록 메인
-	@RequestMapping(value = "shopAdd.admin", method = RequestMethod.GET)
-	public String shopAdd() {											
-		return "admin/shop/shopAdd";
-	}
-		
-	//상품등록 하기
-	@RequestMapping(value = "shopAdd.admin", method = RequestMethod.POST)
-	public void shopAdd2(ShopAddVO vo,HttpServletResponse response,MultipartHttpServletRequest mtf) throws IOException {	
-		System.out.println("mtf :"+mtf);
-		String fileTag = "file";
-		String filePath = "D:\\imageTest\\";
-		
-		MultipartFile file =  mtf.getFile("MAIN_IMG");
-		System.out.println("file : "+file);
-		
-		String fileName= file.getOriginalFilename();
-		
-		file.transferTo(new File(filePath+fileName));
-		
-		System.out.println(vo);
-		response.setContentType("text/html;charset=utf-8");	
-		PrintWriter out = response.getWriter();
-		out.println("<script>");
-//		if (result == 1) {						
-			out.println("alert('완료되었습니다');");
-			out.println("location.href='selectAll.admin';");
-			out.println("</script>");
-//
-//		}else {
-//			out.println("alert('가입 실패');");
-//			out.println("history.back();");
-//
-//		}
-//		out.println("</script>");
-		out.close();
-	}
-	
-
-		//상품목록
-		@RequestMapping(value = "shopList.admin", method = RequestMethod.GET)
-		public String shopList() {											
-			return "admin/member/list";
-		}
-		
-		
-		
 		//테스트용
 		@RequestMapping(value = "shop.dddddddaaan", method = RequestMethod.GET)
 		public String Tesddddt() {
