+++ shop/src/main/java/com/shop/GH/dao/admin/ShopDAO.java
@@ -1,0 +1,5 @@
+package com.shop.GH.dao.admin;
+
+public interface ShopDAO {
+
+}
