+++ shop/src/main/java/com/shop/GH/dao/admin/ShopDAOImpl.java
@@ -1,0 +1,14 @@
+package com.shop.GH.dao.admin;
+
+import javax.inject.Inject;
+
+import org.mybatis.spring.SqlSessionTemplate;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class ShopDAOImpl implements ShopDAO{
+	
+	@Inject
+	private SqlSessionTemplate sql;
+
+}
