+++ shop/src/main/java/com/shop/GH/service/admin/ShopService.java
@@ -1,0 +1,5 @@
+package com.shop.GH.service.admin;
+
+public interface ShopService {
+
+}
