+++ shop/src/main/webapp/WEB-INF/views/admin/home.jsp
@@ -1,13 +1,22 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8"%>
+	pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
+
 <!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Insert title here</title>
-</head>
+<html class="no-js" lang="zxx">
+<jsp:include page="../include_main/header_css.jsp" />
 <body>
-<h3>여기는 관리자페이지 기본페이지</h3>
-<a href="selectAll">모든회원조회</a>
+
+	<div class="wrapper">
+		<jsp:include page="../include_main/header.jsp" />
+		<div class="banner-area hm1-banner pt-130 pb-107"><!-- divnanner -->
+		         
+            </div><!-- divbannerEnd -->
+
+		<jsp:include page="../include_main/footer.jsp" />
+		
+	</div><!-- wrapperEnd -->
+	<jsp:include page="../include_main/footer_js.jsp" />
+
 </body>
 </html>
