+++ shop/src/main/webapp/WEB-INF/views/admin/shop/shopAdd.jsp
@@ -55,6 +55,8 @@
 				</form>
 
 			</div>
+			
+			<h3><%=request.getRealPath("/") %></h3>
 		</div>
 		<!-- divbannerEnd -->
